Download the NASA Soil Moisture Active Passive (SMAP) data for Montana

This is a wrapper for smapr::find_smap() and smapr::download_smap() that crops the resulting data to a ten kilometer buffer around the state of Montana and allows the user to specify "latest" for the desired dates, returning the most recent SMAP product.

mco_get_smap(id = "SPL4SMGP", dates = "latest", raw_dir = NULL)

Arguments

id
A character vector of SMAP dataset to extract.

SPL3FTA

Radar Northern Hemisphere Daily Freeze/Thaw State

SPL3SMA

Radar Global Daily Soil Moisture

SPL3SMP

Radiometer Global Soil Moisture

SPL3SMAP

Radar/Radiometer Global Soil Moisture

SPL4SMAU

Surface/Rootzone Soil Moisture Analysis Update

SPL4SMGP

Surface/Rootzone Soil Moisture Geophysical Data

SPL4SMLM

Surface/Rootzone Soil Moisture Land Model Constants

SPL4CMDL

Carbon Net Ecosystem Exchange

dates
An object of class Date or a character string formatted as To search for one specific date, this can be a Date object of length one. To search over a time interval, it can be a multi-element object of class Date such as produced by seq.Date. Use "latest" (the default) to retreive the most recent SMAP product.

raw_dir
A directory in which to download the raw SMAP data. If missing, data are downloaded to a temporary directory.

Value

A raster brick of the desired SMAP dataset.

Examples

# NOT RUN {
mco_get_smap()
# }