Download and process the Montana SNOTEL Snow Water Equivalent dataset from the NRCS National Water and Climate Center

Download and process the Montana SNOTEL Snow Water Equivalent dataset from the NRCS National Water and Climate Center

mco_get_swe_basins(date = "latest", huc = 6, min_stations = 3)

Arguments

date
A character string of either the date (in yyyy-mm-dd format) or "latest" to get the latest records. Defaults to "latest".

huc
An integer indicating the Watershed Boundary Hydrological Unit level, either "6" or "8", over which to aggregate station data. Defaults to 6.

min_stations
An integer indicating the minimum number of stations that must be available in a region in order to calculate a regional average SWE. Defaults to 3.

Value

A simple feature collection with 9 fields:

WBD code — the WBD identifier

Watershed — the WBD watershed name

Stations — the count of stations aggregated to generate a value for the watershed

SWE (in) — the Snow Water Equivalent (in) start of day values

SWE 1981-2010 Median (in) — the median Snow Water Equivalent (1981-2010) (in) start of day values

Percent SWE — SWE (in) / SWE 1981-2010 Median (in)

Examples

mco_get_swe_basins()

#> Parsed with column specification:
#> cols(
#>   station = col_character(),
#>   `1` = col_integer(),
#>   `2` = col_integer(),
#>   `3` = col_integer(),
#>   `4` = col_integer(),
#>   `5` = col_integer(),
#>   `6` = col_integer()
#> )

#> Parsed with column specification:
#> cols(
#>   `Station Id` = col_integer(),
#>   `Snow Water Equivalent (in) Start of Day Values` = col_double(),
#>   `Median Snow Water Equivalent (1981-2010) (in) Start of Day Values` = col_double()
#> )

#> Joining, by = "WBD code"

#> Warning: attribute variables are assumed to be spatially constant throughout all geometries

#> Simple feature collection with 11 features and 6 fields
#> geometry type:  GEOMETRY
#> dimension:      XY
#> bbox:           xmin: 111853.5 ymin: 17706.01 xmax: 976831.2 ymax: 547983.7
#> epsg (SRID):    102300
#> proj4string:    +proj=lcc +lat_1=45 +lat_2=49 +lat_0=44.25 +lon_0=-109.5 +x_0=600000 +y_0=0 +ellps=GRS80 +towgs84=0,0,0,0,0,0,0 +units=m +no_defs
#> # A tibble: 11 x 7
#>    `WBD code` Watershed       `Stations Count` `SWE (in)` `SWE 1981-2010 Media…
#>    <chr>      <fct>                      <int>      <dbl>                 <dbl>
#>  1 100200     Missouri Headw…               45      575                   484  
#>  2 100301     Upper Missouri                16      226                   146  
#>  3 100302     Marias                         4       64.6                  40.4
#>  4 100401     Fort Peck Lake                 6       61.6                  42.8
#>  5 100402     Musselshell                    7       86.7                  58.8
#>  6 100700     Upper Yellowst…               34      724                   423  
#>  7 100800     Big Horn                      36      417                   295  
#>  8 100901     Tongue                         6       66.3                  55.3
#>  9 100902     Powder                        10       91.8                  62.0
#> 10 170101     Kootenai                      13      279                   236  
#> 11 170102     Pend Oreille                  70     1509                  1126  
#> # ... with 2 more variables: `Percent SWE` <dbl>, geometry <sf_geometry [m]>